Moved lookups logic to separate file
